x:fix: Correct config env var precedence; refactor deepmerge to utils
This commit addresses a failing test and improves code organization.
- Corrects `from_yaml` logic in `config.py` to ensure environment variables take precedence over YAML file settings.
- Moves the `deepmerge` utility function from `ant31box/importer.py` to a new `ant31box/utils.py` module.
- Updates `ant31box/config.py` to import `deepmerge` from the new `utils` module.
