+++ ant31box/importer.py
@@ -27,28 +27,8 @@
 # CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-import collections.abc
 import importlib
 from typing import Any
-
-
-def deepmerge(source: dict, destination: dict) -> dict:
-    """
-    Deeply merges two dictionaries.
-
-    Args:
-        source: The source dictionary.
-        destination: The destination dictionary.
-
-    Returns:
-        The merged dictionary.
-    """
-    for key, value in source.items():
-        if isinstance(value, collections.abc.Mapping):
-            destination[key] = deepmerge(value, destination.get(key, {}))
-        else:
-            destination[key] = value
-    return destination
 
 
 class ImportFromStringError(Exception):
